session-3-ai-agents/agents/holiday-planner/requirements.txt
# Holiday Planner Agent - Dependencies

# Google Gemini SDK
google-genai>=1.0.0

# Environment management
python-dotenv>=1.0.0

# Web frameworks
flask>=3.0.0
fastapi>=0.115.0
uvicorn>=0.32.0

# Database
chromadb>=0.5.0

# HTTP client
httpx>=0.27.0
aiohttp>=3.9.0

# Data validation
pydantic>=2.9.0

# Utilities
python-dateutil>=2.9.0
